% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{chembl_compound_moa}
\alias{chembl_compound_moa}
\title{Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return
its mechanism of action (including disease efficacy, trial phase etc)}
\usage{
chembl_compound_moa(
  chembl_ws_base_url = "https://www.ebi.ac.uk/chembl/api/data",
  molecule_chembl_id = NA
)
}
\arguments{
\item{molecule_chembl_id}{ChEMBL molecule identifier}
}
\value{
a data frame with the following columns of chemical compound properties:
\itemize{
\item molecule_chembl_id_salt
\item molecule_chembl_id = parent_chembl_id,
\item target_chembl_id
\item chembl_disease_efficacy
\item chembl_max_phase
\item chembl_mechanism_comment
\item chembl_moa
}
}
\description{
Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return
its mechanism of action (including disease efficacy, trial phase etc)
}
\examples{
## Retrieve mechanism of action for IRINOTECAN (ChEMBL molecular compound ID = 'CHEMBL481')
\dontrun{
chembl_compound_moa(molecule_chembl_id = 'CHEMBL481')
}

}
