% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{chembl_compound_properties}
\alias{chembl_compound_properties}
\title{Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return a range of compound properties}
\usage{
chembl_compound_properties(
  chembl_ws_base_url = "https://www.ebi.ac.uk/chembl/api/data",
  molecule_chembl_id = NA
)
}
\arguments{
\item{molecule_chembl_id}{ChEMBL molecule identifier}
}
\value{
a data frame with the following columns of chemical compound properties:
\itemize{
\item molecule_chembl_id
\item parent_chembl_id
\item chembl_acd_logp
\item chembl_acd_logd
\item chembl_acd_most_apka
\item chembl_acd_most_bpka
\item chembl_alogp
\item chembl_aromatic_rings
\item chembl_full_molformula
\item chembl_full_mwt
\item chembl_hba
\item chembl_hba_lipinski
\item chembl_hbd
\item chembl_hbd_lipinski
\item chembl_heavy_atoms
\item chembl_molecular_species
\item chembl_mw_freebase
\item chembl_mw_monoisotopic
\item chembl_num_alerts
\item chembl_num_lipinski_ro5_violations
\item chembl_num_ro5_violations
\item chembl_psa
\item chembl_qed_weighted
\item chembl_ro3_pass
\item chembl_rtb
\item chembl_canonical_smiles
\item chembl_standard_inchi
\item chembl_standard_inchi_key
\item chembl_pref_name
\item lincs_id
\item drugbank_id
\item pharmgkb_id
}
}
\description{
Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return a range of compound properties
}
\examples{
## Retrieve compound properties for IRINOTECAN (ChEMBL molecular compound ID = 'CHEMBL481')
\dontrun{
chembl_compound_props(molecule_chembl_id = 'CHEMBL481')
}

}
