% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{chembl_compound_salts}
\alias{chembl_compound_salts}
\title{Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return
relationships between molecule parents and salts (https://www.ebi.ac.uk/chembl/ws)}
\usage{
chembl_compound_salts(
  chembl_ws_base_url = "https://www.ebi.ac.uk/chembl/api/data",
  molecule_chembl_id = NA
)
}
\arguments{
\item{molecule_chembl_id}{ChEMBL molecule identifier}
}
\value{
a data frame with the following columns of chemical compound properties:
\itemize{
\item is_parent - logical indicator if a salt is considered a parent compound
\item molecule_chembl_id - query compound identifier
\item parent_chembl_id - parent compound identifier (salt)
}
}
\description{
Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return
relationships between molecule parents and salts (https://www.ebi.ac.uk/chembl/ws)
}
