% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{chembl_compound_target_predictions}
\alias{chembl_compound_target_predictions}
\title{Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return
targets for which there is a predictied binding of a the molecular target (https://www.ebi.ac.uk/chembl/ws)}
\usage{
chembl_compound_target_predictions(
  chembl_ws_base_url = "https://www.ebi.ac.uk/chembl/api/data",
  min_probability = 0.9,
  molecule_chembl_id = NA
)
}
\arguments{
\item{min_probability}{minimum probability of compound-target binding}

\item{molecule_chembl_id}{ChEMBL molecule identifier}
}
\value{
a data frame with the following columns of chemical compound properties:
\itemize{
\item target_chembl_id
\item target_uniprot_acc
\item molecule_chembl_id_salt
\item molecule_chembl_id
\item target_uniprot_id
\item symbol
\item entrezgene
\item gene_name
\item target_compound_binding_probability
}
}
\description{
Function that accepts a ChEMBL molecule identifer and uses the ChEMBL web service API to return
targets for which there is a predictied binding of a the molecular target (https://www.ebi.ac.uk/chembl/ws)
}
\examples{
## Retrieve predicted targets for IRINOTECAN (ChEMBL molecular compound ID = 'CHEMBL481')
\dontrun{
chembl_compound_target_predictions(molecule_chembl_id = 'CHEMBL481', probability = 0.9)
}

}
