% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compound_synonyms}
\alias{compound_synonyms}
\title{Dataset of compound synonyms - linking known drug aliases to a primary identifier (ChEMBL)}
\format{
\bold{compound_synonyms} - A data frame with 155,940 rows and 3 columns:
\itemize{
\item \emph{alias} - drug alias
\item \emph{nci_concept_display_name} - Primary drug name (NCI Thesaurus)
\item \emph{molecule_chembl_id} - ChEMBL compound identifier
}
}
\source{
\url{https://targetvalidation.org/}

\url{https://www.ebi.ac.uk/chembl/}

\url{https://ncithesaurus.nci.nih.gov/ncitbrowser/}
}
\usage{
compound_synonyms
}
\description{
Dataset of compound synonyms - linking known drug aliases to a primary identifier (ChEMBL)
}
\keyword{datasets}
