% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{get_compound_properties}
\alias{get_compound_properties}
\title{Function that uses the DGIdb (https://dgidb.org) rescource to return all reported compound-target interactions for a given set of human protein targets}
\usage{
get_compound_properties(
  molecule_chembl_id = NULL,
  pchem_cid = NA,
  version_chembl = "ChEMBL_27",
  version_dgidb = "4.0.2"
)
}
\arguments{
\item{genes}{Comma-separated list of targets (gene symbol)}
}
\value{
a data frame with the following columns of chemical compound properties:
\itemize{
\item symbol - Gene/protein symbol
\item gene_name - Gene/protein description
\item dgidb_source - Drug databases/resources (DrugBank, TTD, ChEMBL etc) that indicate support for the compound-target interaction
\item dgidb_interaction_type - Type of drug interaction (agonist, inihibitor etc.)
\item dgidb_pmid - PubMed IDs that support the reported interaction
\item dgidb_interaction_id - DGIdb identifier for the interaction
\item dgidb_gene_categories - Supported gene categories
\item molecule_chembl_id - ChEMBL molecule (compound) ID
\item dgidb_drug_name - Drug name from DGIdb
\item dgidb_score - Confidence for reported interaction (Number of sources (dgidb_source) + Number of PMIDs (dgidb_pmid))
\item pubchem_cid <- PubChem compound identifier (cid)
}
}
\description{
Function that uses the DGIdb (https://dgidb.org) rescource to return all reported compound-target interactions for a given set of human protein targets
}
\examples{
## Retrieve drug interactions for PARP1 + PARP2
\dontrun{
dgidb_drug_interactions(targets = "PARP1,PARP2")
}

}
