% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_drug_records}
\alias{get_drug_records}
\title{Function that retrieves pharmaOncoX data from Google Drive}
\usage{
get_drug_records(cache_dir = NA, force_download = F)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should force downloaded
(i.e. set to TRUE to re-download even if data exists in cache)}
}
\description{
Function that retrieves pharmaOncoX data from Google Drive
}
\keyword{internal}
