% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharm_oncox.R
\name{get_drugs}
\alias{get_drugs}
\title{Get antineoplastic drugs and drug regimens}
\usage{
get_drugs(
  cache_dir = NA,
  force_download = FALSE,
  exclude_salt_forms = TRUE,
  exclude_adc = FALSE,
  treatment_category = c("targeted_therapy_classified", "targeted_therapy_unclassified",
    "chemo_therapy_classified", "hormone_therapy_classified",
    "immuno_suppressants_classified", "other"),
  drug_is_approved = FALSE,
  drug_target = NULL,
  drug_action_type = NULL,
  drug_indication_main = NULL,
  drug_source_opentargets = FALSE,
  drug_cancer_indication = TRUE,
  drug_classified_cancer = TRUE,
  drug_has_blackbox_warning = FALSE,
  drug_approval_year = 1939,
  drug_minimum_phase_any_indication = 0,
  output_resolution = "drug2target2indication",
  drug_action_inhibition = F
)
}
\arguments{
\item{cache_dir}{local cache directory for data retrieval}

\item{force_download}{force download data from remote repository even if
data exists in cache}

\item{exclude_salt_forms}{exclude salt forms of drugs}

\item{exclude_adc}{exclude antibody-drug conjugates (ADCs)}

\item{treatment_category}{main treatment category, classified according to
ATC or not ('targeted_therapy_classified',
'targeted_therapy_unclassified','chemo_therapy_classified','hormone_therapy_classified',
'immuno_suppressants_classified','other')}

\item{drug_is_approved}{logical indicating if resulting drug records
should contain approved drugs only}

\item{drug_target}{character vector with drug targets (gene symbols)
for drug records included in results}

\item{drug_action_type}{character vector with drug action types to
include in drug record list -
possible values "INHIBITOR","AGONIST","MODULATOR","ANTAGONIST",
"BLOCKER","ACTIVATOR","BINDING AGENT","OPENER",
"STABILISER","CROSS-LINKING AGENT",DISRUPTING AGENT","OTHER"}

\item{drug_indication_main}{character vector with main tumor types for
which drug(s) are indicated. Possible values: "Adrenal Gland","Biliary Tract",
"Bladder/Urinary Tract","Bone","Breast","Cervix","CNS/Brain",
"Colon/Rectum","Esophagus/Stomach","Eye","Head and Neck","Kidney",
"Liver","Lung","Lymphoid","Myeloid","Ovary/Fallopian Tube",
"Pancreas","Penis","Peripheral Nervous System","Peritoneum",
"Pleura","Prostate","Skin","Soft Tissue","Testis","Thymus",
"Thyroid","Uterus","Vulva/Vagina"}

\item{drug_source_opentargets}{logical indicating if resulting drug records
should contain drug records from Open Targets Platform/ChEMBL only}

\item{drug_cancer_indication}{logical indicating if resulting drug records
should be for those indicated for cancer conditions only (from approved
conditions, found in clinical trials etc.)}

\item{drug_classified_cancer}{logical indicating if resulting drug
records should be for those classified only in the "L" class of ATC (
"ANTINEOPLASTIC AND IMMUNOMODULATING AGENTS") only}

\item{drug_has_blackbox_warning}{logical indicating if resulting drug
records should contain drugs with black box warnings only}

\item{drug_approval_year}{only include records for drugs approved
later than this date (year)}

\item{drug_minimum_phase_any_indication}{only include drug records that
are in a clinical phase (any indication) greater or equal than this phase}

\item{output_resolution}{dictate output record resolution
('drug','drug2target','drug2target2indication')}

\item{drug_action_inhibition}{logical indicating to only return drug records
with inhibitory mechanism-of-action}
}
\value{
The \code{records} data frame contains the following columns (only selected columns
will be shown based on the value of \code{output_resolution})

\itemize{
\item \emph{drug_id} - drug identifier (pharmaOncoX)
\item \emph{drug_name} - primary drug name (upper case, NCI Thesaurus)
\item \emph{drug_type} - type of drug molecule (Antibody, small molecule etc)
\item \emph{molecule_chembl_id} - ChEMBL compound identifier
\item \emph{drug_action_type} - main action elicited by drug
(antagonist, inhibitor, stabiliser etc)
\item \emph{drug_alias} - collection of unambiguous drug aliases
(separated by '|')
\item \emph{nci_concept_definition} - detailed description of drug
mechanism-of-action (NCI Thesaurus)
\item \emph{opentargets} - logical - drug is found in the Open Targets
Platform resource
\item \emph{is_salt} - logical - drug record represents a salt form
(excluded by default)
\item \emph{is_adc} - logical - drug record represents an
antibody-drug conjugate (ADC - excluded by default)
\item \emph{drug_blacbox_warning} - logical indicating if drug has
blackbox warning
\item \emph{nci_t} - NCI thesaurus identifier
\item \emph{target_symbol} - gene symbol of drug target
\item \emph{target_entrezgene} - Entrez gene identifier of drug target
\item \emph{target_genename} - gene name/description of drug target
\item \emph{target_ensembl_gene_id} - Ensembl gene identifier of drug target
\item \emph{target_type} - type of drug target (single protein, protein
family etc.)
\item \emph{drug_max_phase_indication} - maximum clinical phase for drug
(given indication)
\item \emph{drug_approved_indication} - logical indicating if drug has
an approved indication
\item \emph{drug_frac_cancer_indications} - fraction of drug indications that
are for cancers
\item \emph{drug_approved_noncancer} - logical indicating if drug
is approved for a non-cancer disease
\item \emph{drug_n_indications} - number of indications for the
given drug (from approved indications, clinical trials etc)
\item \emph{drug_year_first_approval} - year drug was first approved
\item \emph{drug_max_ct_phase} - maximum clinical phase for drug
(any indication)
\item \emph{disease_efo_id} - EFO (Experimental Factor Ontology)
identifier for drug indication
\item \emph{disease_efo_label} - EFO (Experimental Factor Ontology)
label for drug indication
\item \emph{primary_site} - primary tumor site/type (obtained through
https://github.com/sigven/oncoPhenoMap)
\item \emph{drug_clinical_id} - drug clinical identifier
(clinicaltrials.gov, DailyMed, FDA etc.)
\item \emph{drug_clinical_source} - underlying source for drug entry
(DailyMed, clinicaltrials.gov, FDA etc.)
\item \emph{atc_code_level1} - drug identifier ATC (level 1)
\item \emph{atc_level1} - drug label ATC (level 1)
\item \emph{atc_code_level1} - drug identifier ATC (level 2)
\item \emph{atc_level2} - drug label ATC (level 2)
\item \emph{atc_code_level3} - drug identifier ATC (level 3)
\item \emph{atc_level3} - drug label ATC (level 3)
\item \emph{atc_treatment_category} - treatment category (targeted/chemo/hormone, cancer/other etc)
}
}
\description{
Downloads preprocessed datasets to a local cache directory and returns a
selected set of drugs based on various criteria set by the user.

The dataset comes as a \code{list} object, with two elements:
\itemize{
\item \code{metadata} - a data frame with metadata regarding drug resources used
\item \code{records} - a data frame with drug records
}
}
