% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_on_off_label_drugs}
\alias{get_on_off_label_drugs}
\title{Utility function to get on-label/off-label drugs (not complete)}
\usage{
get_on_off_label_drugs(cache_dir = NA)
}
\arguments{
\item{cache_dir}{Local cache directory}
}
\description{
Utility function to get on-label/off-label drugs (not complete)
}
\keyword{internal}
