% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db.R
\name{get_onco_drugs}
\alias{get_onco_drugs}
\title{Retrieve antineoplastic drugs and drug regimens}
\usage{
get_onco_drugs(
  drug_is_targeted = F,
  drug_is_approved = F,
  drug_target = NULL,
  drug_action_type = NULL,
  drug_indication_main = NULL,
  source_opentargets_only = F,
  drug_has_blackbox_warning = F,
  drug_approved_later_than = 1939,
  drug_minimum_phase_any_indication = 0,
  list_per_drug_synonym = F,
  output_resolution = "drug2target2indication",
  output_style = "extensive",
  is_alkylating_agent = F,
  is_angiogenesis_inhibitor = F,
  is_anthracycline = F,
  is_antimetabolite = F,
  is_ar_antagonist = F,
  is_bet_inhibitor = F,
  is_hedgehog_antagonist = F,
  is_hdac_inhibitor = F,
  is_hormone_therapy = F,
  is_immune_checkpoint_inhibitor = F,
  is_kinase_inhibitor = F,
  is_monoclonal_antibody = F,
  is_parp_inhibitor = F,
  is_platinum_compound = F,
  is_proteasome_inhibitor = F,
  is_topoisomerase_inhibitor = F,
  is_tubulin_inhibitor = F
)
}
\arguments{
\item{drug_is_targeted}{logical indicating if resulting drug records should contain molecularly targeted drugs only}

\item{drug_is_approved}{logical indicating if resulting drug records should contain approved drugs only}

\item{drug_target}{character vector with drug targets (gene symbols) for drug records included in results}

\item{drug_action_type}{character vector with drug action types to include in drug record list -
possible values "INHIBITOR","AGONIST","MODULATOR","ANTAGONIST","BLOCKER","ACTIVATOR","BINDING AGENT","OPENER",
"STABILISER","CROSS-LINKING AGENT",DISRUPTING AGENT","OTHER"}

\item{drug_indication_main}{character vector with main tumor types for which drug(s) are indicated
possible values "Adrenal Gland","Biliary Tract","Bladder/Urinary Tract","Bone","Breast","Cervix","CNS/Brain",
"Colon/Rectum","Esophagus/Stomach","Eye","Head and Neck","Kidney","Liver","Lung","Lymphoid","Myeloid","Ovary/Fallopian Tube",
"Pancreas","Penis","Peripheral Nervous System","Peritoneum","Pleura","Prostate","Skin","Soft Tissue","Testis","Thymus",
"Thyroid","Uterus","Vulva/Vagina"}

\item{source_opentargets_only}{logical indicating if resulting drug records should contain drug records from Open Targets Platform only}

\item{drug_has_blackbox_warning}{logical indicating if resulting drug records should contain drugs with black box warnings only}

\item{drug_approved_later_than}{only include records for drugs approved later than this date (year)}

\item{drug_minimum_phase_any_indication}{only include drug records that are in a clinical phase (any indication) greater or equal than this phase}

\item{list_per_drug_synonym}{list resulting drug records per drug synonym}

\item{output_resolution}{output record resolution ('drug','drug2target','drug2target2indication')}

\item{output_style}{comprehensiveness of drug records ('extensive' - all annotations, 'narrow' - main annotations)}

\item{is_alkylating_agent}{logical indicating if only this drug class is wanted}

\item{is_angiogenesis_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_anthracycline}{logical indicating if only this drug class is wanted}

\item{is_antimetabolite}{logical indicating if only this drug class is wanted}

\item{is_ar_antagonist}{logical indicating if only this drug class is wanted}

\item{is_bet_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_hedgehog_antagonist}{logical indicating if only this drug class is wanted}

\item{is_hdac_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_hormone_therapy}{logical indicating if only this drug class is wanted}

\item{is_immune_checkpoint_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_kinase_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_monoclonal_antibody}{logical indicating if only this drug class is wanted}

\item{is_parp_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_platinum_compound}{logical indicating if only this drug class is wanted}

\item{is_proteasome_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_topoisomerase_inhibitor}{logical indicating if only this drug class is wanted}

\item{is_tubulin_inhibitor}{logical indicating if only this drug class is wanted}
}
\description{
Retrieve antineoplastic drugs and drug regimens
}
