% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{get_similar_compounds}
\alias{get_similar_compounds}
\title{Function that accepts a SMILEs and queries PubChem for structurally similar compounds}
\usage{
get_similar_compounds(
  approach = "pubchem_pug",
  smiles = NULL,
  manimoto_threshold = 100,
  MaxRec = 100,
  MaxSec = 10,
  mw = NULL
)
}
\arguments{
\item{approach}{'pubchem_pug' or 'chemminer'}

\item{smiles}{query SMILES}

\item{manimoto_threshold}{minimum similarity}

\item{MaxRec}{maximum number of records}

\item{MaxSec}{maximum number of}
}
\value{
a character vector of PubChem compound identifiers (CIDs)
}
\description{
Function that accepts a SMILEs and queries PubChem for structurally similar compounds
}
