% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharm_oncox.R
\name{match_drug_names}
\alias{match_drug_names}
\title{Match a list of drug names/aliases to the dataset of drugs
provided by pharmOncoX}
\usage{
match_drug_names(
  query = NULL,
  exclude_salt_forms = TRUE,
  exclude_adc = FALSE,
  cache_dir = NA,
  force_download = FALSE
)
}
\arguments{
\item{query}{A character vector of drug names/aliases}

\item{exclude_salt_forms}{Logical indicating if salt forms should be excluded}

\item{exclude_adc}{Logical indicating if antibody-drug conjugates should be excluded}

\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should force downloaded}
}
\description{
Match a list of drug names/aliases to the dataset of drugs
provided by pharmOncoX
}
