% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncopharma_synonyms}
\alias{oncopharma_synonyms}
\title{Dataset of drug synonyms - linking known drug aliases to a primary identifier}
\format{
\bold{oncopharma_synonyms} - A data frame with 9,946 rows and 52 columns:
\itemize{
\item \emph{alias} - drug alias
\item \emph{nci_concept_display_name} - Primary drug name (NCI Thesaurus)
\item \emph{molecule_chembl_id} - ChEMBL compound identifier
}
}
\source{
\url{https://targetvalidation.org/}

\url{https://www.ebi.ac.uk/chembl/}

\url{https://ncithesaurus.nci.nih.gov/ncitbrowser/}
}
\usage{
oncopharma_synonyms
}
\description{
Dataset of drug synonyms - linking known drug aliases to a primary identifier
}
\keyword{datasets}
