% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{pubchem_compound_properties}
\alias{pubchem_compound_properties}
\title{Function that retrieves PubChem compound properties using PubChem public user gateway (PUG)}
\usage{
pubchem_compound_properties(
  PUBCHEM_PUG_URL = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/",
  pubchem_cid = 9444
)
}
\arguments{
\item{pubchem_cid}{PubChem compound identifier}
}
\value{
a data frame with the following columns of chemical compound properties:
\itemize{
\item pubchem_isomeric_SMILES
\item pubchem_TPSA
\item pubchem_IUPAC_name
\item pubchem_complexity
\item pubchem_hbond_donor_count
\item pubchem_hbond_acceptor_count
\item pubchem_molecular_weight
\item pubchem_cid
\item pubchem_name
\item molecule_chembl_id
}
}
\description{
Function that retrieves PubChem compound properties using PubChem public user gateway (PUG)
}
\examples{
## Retrieve compound properties for Azacitidine (Pubchem compound ID = 9444)
\dontrun{
pubchem_compound_props(pubchem_cid = 9444)
}

}
