% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db_utilities.R
\name{readUrl}
\alias{readUrl}
\title{Function that retrieves PubChem compound properties using PubChem public user gateway (PUG)}
\usage{
readUrl(q_url)
}
\arguments{
\item{q_url}{Query URL}
}
\value{
a data frame with the following columns of chemical compound properties:
}
\description{
Function that retrieves PubChem compound properties using PubChem public user gateway (PUG)
}
