% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targets}
\alias{targets}
\title{Dataset of therapeutic and actionable targets in cancer}
\format{
\bold{druggable} - A data frame with 1,094 rows and 6 columns:
\itemize{
\item \emph{symbol} - Gene symbol
\item \emph{ensembl_gene_id} - Ensembl gene identifier
\item \emph{genename} - Gene description
\item \emph{entrezgene} - Gene identifier (Entrez)
\item \emph{nci_concept_display_name} - Primary compound name (separated by '|')
\item \emph{molecule_chembl_id} - ChEMBL compound identifier (separated by '|')
}

\bold{actionable} - A data frame with 623 rows and 7 columns:
\itemize{
\item \emph{symbol} - Gene symbol
\item \emph{ensembl_gene_id} - Ensembl gene identifier
\item \emph{genename} - Gene description
\item \emph{entrezgene} - Gene identifier (Entrez)
\item \emph{phenotypes} - List of associated phenotypes for actionable gene
\item \emph{clinical_significance} - Clinical significance of biomarker (prognosis, drug sensitivity etc)
\item \emph{source_db} - Underlying source database for actionability evidence
}
}
\usage{
targets
}
\description{
A list of two data frames:
\itemize{
\item druggable: data frame that indicates all drugs/compounds pr. target
\item actionable: data frame with actionable genes in cancer (biomarkers ,
for therapies, prognosis etc. (harvested from various sources)
}
}
\keyword{datasets}
