% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharm_oncox.R
\name{get_biomarkers}
\alias{get_biomarkers}
\title{Get curated cancer biomarker datasets}
\usage{
get_biomarkers(cache_dir = NA, force_download = F)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should force downloaded
(i.e. set to TRUE to re-download even if data exists in cache)}
}
\value{
Each entry of the source-specific (e.g. 'civic') entry in the \code{data} list contains
a list of three data frames:

\itemize{
\item \emph{variant} - list of all biomarker variants, extensively populated
according to variant aliases (identifer - column \strong{variant_id})
\item \emph{clinical} - cross-references between variants recorded in
the \code{variant} data frame and clinical evidence items (identifier -
column \strong{evidence_id}) and underlying literature evidence
(identifier - column \strong{source_id})
\item \emph{literature} - lists literature for all source_id's listed in
the \code{clinical} data frame
}
}
\description{
Downloads preprocessed datasets to a local cache directory and returns a
curated set of genomic biomarkers from multiple sources
(CIViC, CGI, MitelmanDB)

The dataset comes as a \code{list} object, with three elements:
\itemize{
\item \code{metadata} - a data frame with metadata regarding drug resources used
\item \code{data} - a list with four elements ('civic','cgi','mitelmandb','custom_fusions')
\item \code{fpath} - path to cache file
}
}
