% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_db.R
\name{get_drug}
\alias{get_drug}
\title{Function that returns various properties of antineoplastic drugs (targeted and non-targeted)}
\usage{
get_drug(
  drug_is_targeted = F,
  drug_is_approved = F,
  drug_target = NULL,
  drug_action_type = NULL,
  source_opentargets_only = F,
  drug_has_blackbox_warning = F,
  drug_approved_later_than = 1939,
  drug_minimum_max_phase_any_indication = 0,
  list_per_drug_only = F,
  list_per_indication = T,
  list_per_drug_synonym = F
)
}
\arguments{
\item{drug_is_targeted}{logical indicating if resulting drug records should contain targeted drugs only}

\item{drug_is_approved}{logical indicating if resulting drug records should contain approved drugs only}

\item{drug_target}{character vector with drug targets (gene symbols) for drug records included in results}

\item{drug_action_type}{character vector with drug action types to include in drug record list -
possible values "INHIBITOR","AGONIST","MODULATOR","ANTAGONIST","BLOCKER","ACTIVATOR","BINDING AGENT","OPENER",
"STABILISER","CROSS-LINKING AGENT",DISRUPTING AGENT","OTHER"}

\item{source_opentargets_only}{logical indicating if resulting drug records should contain drug records from Open Targets Platform only}

\item{drug_has_blackbox_warning}{logical indicating if resulting drug records should contain drugs with black box warnings only}

\item{drug_approved_later_than}{only include records for drugs approved later than this date (year)}

\item{drug_minimum_max_phase_any_indication}{only include drug records that are in a clinical phase (any indication) greater or equal than this phase}

\item{list_per_drug_only}{list resulting drug records per drug only, ignore targets and indications}

\item{list_per_indication}{list resulting drug records per drug indication/tumor type}

\item{list_per_drug_synonym}{list resulting drug records per drug synonym}
}
\description{
Function that returns various properties of antineoplastic drugs (targeted and non-targeted)
}
